Fix issue with SIGHUP being sent to processes with recycled pids
This is a fix for an issue which was reported to the ast-developers
list. The details can be found at
https://www.mail-archive.com/ast-developers@lists.research.att.com/msg01887.html

The fix ensures that the HUP signal is sent to the process group only
if there is some process with the same process group id whose P_DONE is not set,
i.e which hasn't yet completed execution and "wait"ed upon yet.
Fix provided by nobotumo.nakano@oracle.com