Use posix exit code if last command exits due to a signal
If the last command in a script was killed by a signal, ksh kills itself with the same signal. This patch fixes it by using exit() (with status 128 + signal number) instead of kill().
It also fixes the bug where ksh kills itself with a signal if argument to exit command is greater than 256.

Resolves: rhbz#1469624