Tell git to ignore some files and directories
There are files and directories that are artifacts of building and
testing this software. And others that are the result of other actions
such as text editors creating backup files. We don't want those
artifacts accidentally included in a commit and it's annoying to have
`git status` telling me they haven't been staged for a commit.