Update some files to match output of bin/package
Running `bin/package build` updates several scripts in the *bin*
directory because they're missing the current `USAGE_LICENSE`
definition. This is super annoying as every time you do a build `git
status` tells you those files have changed and there are now *.old
versions as well that need to be committed. So commit the updated
scripts to eliminate that noise.