Assume `MAP_ANON` is valid
One of the feature tests involving `mmap()` is dubious since it involves
highly variable timing data that can lead to non-deterministic builds.
Several other feature tests are no longer relevant since, for example, I
can't think of a single OS we care about which supports `mmap()` but not
the `MAP_ANON` flag.

Finally, there is no good reason to even use `mmap()` and implement our
own memory allocator. These days we should be using the OS provided libc
`malloc()` family of functions. And the value of the performance speedup
of using it in the SFIO code is dubious for a command shell. However, I
found that when I disabled it completely it results in more test failures
so this change is limited to the MAP_ANON related configuration.

Fixes #84