Assume math.h functions are available
Bypass the incredibly inefficient and expensive (59 seconds on my
system) process of figuring out which math functions are available. It
does this by using a handcrafted version of the C module the
./src/cmd/ksh93/features/math.sh script creates if it can be built and
linked without error. That should be the case for any modern UNIX system
we care about. Thus avoiding the expensive probing for each function
individually.

Fixes #88