Don't use SOCK_CLOEXEC
The `SOCK_CLOEXEC` symbol is not available on some of the platforms
we want to support such as macOS. It's also wrong if it's missing to
define it with a specific value. Instead just use `fcntl()` to set the
close-on-exec attribute.

Also, remove `SOCK_NONBLOCK` because it too is nonstandard and isn't
actually needed to build ksh.

Fixes #110