Fix something involving nv_putval()
This change is pulled from here:

https://github.com/oracle/solaris-userland/blob/master/components/ksh93/patches/065-CR7110983.patch

It does not break any existing unit tests and looks valid. It is
unfortunate that the change comment associated with that URL is not more
descriptive about what is fixed. I'm merging this because I would like
to make some radical changes to the style of the code which will make
merging such changes more difficult. So I am implicitly assuming the
team maintaining ksh on Solaris knew what they were doing.

Fix for #122