Implement config for clang-format and editors
This creates a config file for the clang-format tool. It also establishes
a .editorconfig file to help establish the proper settings for editing
code in this project.

Finally, it includes one module that was reformatted with clang-format
followed by some manual cleanups to illustrate how this affects the
cod. The latter being necessary because clang-format won't do things like
add missing braces. I verified the resulting object file is unchanged
using the `objdump -d` technique.

First step to fixing #125