Add optional dependency on libexecinfo to Meson
After merging my change to implement code to write a backtrace to stderr
upon receipt of SIGSEGV I noticed it failed to build on FreeBSD 11.1
because the `backtrace_symbols()` function couldn't be found. This adds
a check for the library needed on FreeBSD to the Meson config. I'm not
going to bother with the Nmake config because it looks like we'll be
able to replace it with Meson in the immediate future.