Make it unnecessary to `meson -D b_lundef=false`
A little investigation revealed that it is possible to specify default
project options. This makes it unnecessary for the user type explicitly
add `-D b_lundef=false` when running `meson` on BSD systems.

This change also reformats all the meson.build config files to limit the
length of lines and make the formatting of the configs more consistent.
As well as more consistent with how a typical Python program is
formatted. Which is reasonable since the Meson config language is based
on Python.