Add test timeouts where the default is too short
This avoids the need for the user to specify a test timeout multiplier
(e.g., `meson test -t 6`) unless running the tests under the control of
a tool like valgrind.

Don't use the `-v` flag when running `meson test`. Instead display the
full build log if the tests fail. This avoids the problem with the
stdout and stderr streams not being interleaved.

Remove `vim` as an install requirement of the OpenSUSE environment. I
have no idea why it is automatically included in the Fedora environment
but it's not needed so don't explicitly include it.