Reinstate ignoring src/lib/libast/features files
Commit a2499fe removed *src/lib/libast/features* from the *.gitignore*
file because some files in that directory are static which means those
files should be tracked by git. Unfortunately, the legacy Nmake/iffe
code we still depend on puts dynamically generated files in that
directory.  Which makes `git status` report files that should not be
committed.

Ultimately we will stop creating files in that directory once we have
stopped using the `iffe` feature test command. Until then stop telling
developers that there are files in that directory which are not in
the repo.