Enable some features from the legacy build system
While working on issue #210 I noticed there were some features enabled
by the legacy Nmake based build system that were not enabled in the new
Meson based build system. Since that appears to be an oversight enable
those features so that we have parity with the legacy ksh93u+ build.
These symbols include SHOPT_POLL, SHOPT_FILESCAN, and SHOPT_EDPREDICT.

Whether or not the features controlled by those symbols should be in the
next stable release is TBD.

This also refactors the Meson build config to make it more obvious which
features are shared between the ksh and shcomp build targets.