Remove unneeded libast code
A recent change removed the `iffe` based test for some features related
to reading and parsing the mount table file and using related syscalls
like `getmntent()`. Unfortunately the resulting static header only works
on Linux. Fortunately, that header and the code that depends on it is
not needed to build ksh. So remove those files in order that we can once
again build ksh on macOS and FreeBSD.