Correctly detect if `utimensat()` is available
This corrects the feature detection for the `utimensat()` and
`utimets()` functions by including the correct system header.

There is still a bug present in the case where those two functions
aren't available and `tvtouch()` is called with the `av` parameter set
to NULL. But that is a problem for another day.

Fixes #249