Revert part of "Remove unused functions."
This reverts the portion of commit 512c3c5e that removes the `nativeconf()`
function that is needed on Linux systems. But it puts the code under the
same `#if` preprocessor directive as the call site so it won't result in
compiler warnings about an unused function on platforms that don't need it.