Fix for unit test locale.sh and siblings
This PR fixes a "bug" with `join -t`.
Join is a utility that in this particular unit test is used to combine two files satisfying certain criteria.
It turns out that `join -t` accepts a single character or a string of blank or comma separated characters.
The test instead uses a three-character string containing no spaces or blanks.
It aims to use the at-sign (@) as tab character.

The PR changes the input files to use the intended tab character.
It fixes issue #301.