Fix proper indentation
gcc complains that the block inside the for loop is not properly indented.
This PR add another indentation to avoid the warning.
This ia a good example where extra curly braces would both help humands a compilers.
If this change collides with the styling compaign, feel free to reject it.