Ensure the compiler supports C99 features
This code might be compiled on systems old enough that C99 (or newer)
is not the default language standard. This ensures those old gcc
implmentations have C99 support enabled. And if the gcc is so old it
doesn't support that language standard then this will cause the compilation
to fail without even trying to compile the code.

Fixes #145