Restyle modules already restyled based on new conf
The original clang-format config allowed slightly overlong lines. I
decided that was more annoying and confusing than it was worth. So the
previous commit effectively makes overly long lines impossible. This
applies that to the files I had already restyled. This also fixes
several style problems not caught by manual changes to the code that
were not run through clang-format.