Remove the dependency on bash
This removes any use of `bash` features in the build bootstrap scripts.
This allows us to use the legacy Bourne shell, `/bin/sh`, to bootstrap
the build.  Also, since these scripts are not `source`d there is no reason
to use mechanisms like the `pushd` command so just use `cd`.

It also removes the remaining use of `iffe` to detect whether the
headers `stdint.h`, `signal.h`, `stddef.h` and `wchar.h` are available.

This makes is easier to build ksh on systems like OpenBSD which might
not have `bash` installed.