Add a *CHANGELOG.md* file
Most projects have something analoguous to the *CHANGELOG.md* file
in this commit. The intent is to record significant changes from one
stable release to the next. This initial commit is going to be incomplete
because it was created long after the AT&T AST project was open sourced.
But hopefully it captures the significant changes made when maintenance
of the project resumed.