Fix the `!tp->lst.lstlef->tre.tretyp == TCOM` test
This corrects the obviously wrong `tp->lst.lstlef->tre.tretyp == TCOM`
test in *src/cmd/ksh93/sh/name.c*. I did not mask it with `& COMMSK` as
is done in roughly half the places where such comparisons are performed
since the rules for when the type must be masked before being compared
and when it shouldn't is unclear.

I also fixed a couple of instances which caught my eye where the magic
value zero was used instead of the equivalent `TCOM` symbol.

Resolves #338