Fix bogosity involving symbol `_SH_PRIVATE`
The structure members defined by `_SH_PRIVATE` aren't really private.
And defining them in the current fashion makes it harder to understand
the contents of `struct Shell_s`. I'd like to do more refactoring of
this to make the distinction between private and public headers clearer
but this will do for now.