Change temp folder pattern to indicate which test script it is created for
This PR addresses the difficulty to associate the temp folder created by the test scripts.
The temp folder names are unique to allow for repeated invocation, and for the invocation
to happen in parallel.
This PR adds a prefix "ksh." plus the test name in front of the temp folder name.