Fix the incorrect `#ifdef` tests
From comment https://github.com/att/ast/issues/383#issuecomment-360362040:

I started by extracting all the #if directives from the source. What
I noticed is that there are some files which incorrectly do `#ifdef
some_symbol`. Which means if you pass `-Dsome_symbol=0` the wrong thing
happens. The idiom in 98% of the AST code is `#if some_symbol` to test if
it is set true. Yes, the behavior of the legacy iffe tool is apparently
to never emit `#define some_symbol 0`. It only emits the symbol if it
is defined to one (true). But since most of the conditional compilation
works if you explicitly define the symbol to zero (false) the few places
which don't follow that pattern represents a land mine.