Remove several symbols related to locale support
The code doesn't compile if `iconv_open()` and related functions aren't
available. As well as header `locale.h`. Not too that `_lib_locale` was
hardcoded to one so there wasn't even a platform feature test checking
that it was available. And since it's now 2018 we can reasonably expect
every platforme to profide a locale subsystem.