Call exit(3) instead of _exit(2) in posix_spawn feature test
The exit(3) function will flush stdout before terminating the program,
which leads to more consistent output from the printf() calls.

With meson, stdout is redirected to meson-logs/meson-log.txt which means
that output is buffered in large chunks instead of being auto-flushed on
newlines when stdout is attached to a tty device.