Remove unused feature test symbols
I scanned the code for every metion of a symbol defined by a platform
feature test. Over half the symbols in config_ast.h.in are no longer
present in the project as a consequence of pruning code not needed to
build ksh. So this change removes those symbols from the config header.