Enable malloc debug features
This enables malloc debugging features provided by various
implementations. This is only really useful when issue #396 is resolved
so that `ksh` no longer uses the AST Vmalloc subsystem. I'm recommending
merging this now to to make it obvious that any bugs involving the use
of the native malloc subsystem are caught sooner rather than later.