Simplify feature detection
Simplify the feature detection logic. This has the benefit of ensuring
the binary symbols are set to zero if the feature wasn't detected.
This makes it easier to figure out which symbols still need detection
logic by examining the build/config_ast.h file. I also removed a couple
of redundant symbols. For example, if `_lib_gethostname` is defined it
is safe to assume `_lib_sethostname` would have been defined. Thus the
latter symbol is unnecessary.