Fix `locale` test on systems with broken `join`
The `locale` unit test contains a test meant to test the builtin `join`
command. At the moment we don't have that specific builtin enabled. And
the external `join` command on many distros is not Unicode enabled. So
disable that subtest for now. It will be re-enabled if and when those
builtins are reinstated.

Resolves #411