Fix job control if ksh is started as login shell
While reformatting code in 00d0a53b86c5bb9d94c2d1cca74688d6cf67835e,
this code :

  if(possible = (setpgid(0,job.mypgid)>=0) || errno==EPERM)

was changed to:

  possible = setpgid(0, job.mypgid) >= 0;
  if (possible || errno == EPERM) {

So possible is not set to 1 if setpgid() fails due to EPERM. According
to POSIX, calling setpgid() on a session leader should fail with EPERM.
A login shell is always started as a session leader. This caused ksh to
skip initializing job control if it is started as login shell (session
leader).

It should have been changed to:

  possible = (setpgid(0, job.mypgid) >= 0)  || errno == EPERM;
  if (possible) {

This commit fixes it.

Resolves: #437