Remove dead code related to initializing job control
Ksh tries to verify if job control is set up properly by calling
setpgid() on it's own process.

The code that is being removed is taking these steps:

1. job.mypgid = getpgrp(); sets job.mypgid to current process group.
According to POSIX, getpgrp() should never fail. This means job.mypgid
will always be set to a value greater than or equal to 1. So this check
is unnecessary:

    if (job.mypgid <= 0) {

and the code under it will never be executed.

2. possible = setpgid(0, job.mypgid) >= 0 tries to set process group id
of current process to the same value that it got in step 1. It is
unneeded.

We should be able to initalize job control by safely assuming that ksh is
started in a valid process group.

Related: #437