Fix a regression in appending to arrays
Existing elements are getting overwritten while appending to arrays if
an array contains empty indexes. For e.g.

 typeset -a a=([1]=w [2]=x)
 typeset -a b=(a b c)
 a+=("${b[@]}")
 typeset -p a

Actual output:

 typeset -a a=([1]=w [2]=a [3]=b [4]=c)

Expected output:

 typeset -a a=([1]=w [2]=x [3]=a [4]=b [5]=c)

Current code was calculating index of the last element by counting
number of elements in the array. It will not work if an array contains
empty indexes. This commit fixes it.

Resolves: #23