Unify the unit test framework
This modifies how ksh unit tests are run to have a consistent
environment (i.e., preamble) and post test behavior (i.e., postscript).
There was at least one unit test that did not correctly report the
number of errors via the exit status.

Each test is now run in a unique temporary directory. If the test
doesn't complete without error the temp dir is not removed. This makes
it easier when running the tests on a personal system (rather than, say,
a CI environment like Travis) to explore the state of the unit test.

The `HOME` env var is now set to the temp dir to help keep the test from
modifying the user's environment (e.g., their ~/.sh_history file).

Resolves #429
Resolves #477