Check for invalid file descriptor in sh_close()
There was a regression introduced after the last stable release which
causes mishandling of redirections in subshells. For e.g.

foo=$(print `/bin/echo bar`)
print foo=$foo

prints `bar` on stdout and the variable foo is empty.

Handling of redirections in $() style of command substitution was
changed after the last stable release. sh_subshell() function replaces
sfstdout with a newly created temporary stream with this line :

sp->saveout = sfswap(sfstdout, NULL);

The newly created stream has it's fd value set to -1. If `` style of
command subtitution happens inside $(), it tries to use this stream with
invalid fd and fails to close it. So redirections for stdout do not
work.

We should check if the stream that we are closing actually corresponds
to the file descriptor that we want to close. If not, directly invoke
close() on the fd.

Resolves: #478