Replace $tmp with $TEST_DIR
The existing unit tests are confusing due to their uses of `$tmp` to
refer to the temporary directory created for the unit test. While at the
same time using that symbol, and using it as a prefix for other symbols,
that are unrelated or only indirectly related to the temp dir. So replace
those uses with `$TEST_DIR`.