Keep exit status consistent when command is not executable
Ksh performs command searching by trying to execute command in each
element of PATH. If a command is found, but is not executable, ksh
continues to search for it further in other paths.

According to POSIX, if shell finds a command which is not executable,
it should return with status 126. Current behavior of ksh is non
determinstic as it depends on if command is found in the last element
of PATH or not:

1. If a command is found in the last element of PATH but is not
executable, ksh returns with status 126.

2. If a command is found before last element of PATH but is not
executable, ksh continues to search  for it further in other paths. It
returns with status 127 when command could not be found.

Ksh should behave consistently and return immediately after a command is
found but is not executable.

Resolves: #485