In case statement, stop trying pattern as literal string
`case "[0-9]" in [0-9]) echo match;; esac` now stops matching. When a
case statement doesn't match a pattern -- as the string "[0-9]"
doesn't match the pattern [0-9] for a single digit -- then it no
longer tries to use the pattern as a literal string.

Now `case word in pattern)` in ksh behaves like `[[ word = pattern ]]`
in ksh and case statements in most other shells.

Resolves #476