Fix pathdev() handling of symlinks
This fixes the io.sh unit test on any platform which has two consecutive
symlinks in the temp directory pathname that points to a path longer than
the symlink. It also introduces a regression test which fails on Linux
(where /tmp isn't normally a symlink) without this change.

Note that I do not claim this change is 100% correct. The original code
is borderline impossible to understand. But this does fix the problem
and doesn't introduce any new unit test failures.

Partial fix for #483
Resolves #492