Check for systems where "x" is a known locale
OpenBSD accepts "x" as an alias of "C", and I can't use `locale -a` to
detect this.  If $(uname) is OpenBSD, then use "x.x" as the unknown
locale.

Most other systems agree that "x" is unknown.  If `locale -a` shows
that someone added the "x" locale, then skip the test.

Partial fix for #483