Remove support for SHOPT_ACCTFILE
The SHOPT_ACCTFILE feature is not enabled by default. As far as I can
tell it isn't enabled in any mainstream distro. The top Google search
result is

https://blog.fpmurphy.com/2008/12/ksh93-auditing-and-accounting.html

The feature is a security nightmare. Which is probably why it isn't
enabled by default or any distro that includes ksh. It should never have
been added. It certainly has no business being enabled today given all
the hackers looking for ways to cause mischief.

This is related to issue #210 and PR #214.