Stop creating the conftab files in the source tree
The legacy (IFFE based) build system generated several files in the
source tree. That is not a best practice since it requires, among other
things, that the VCS be configured to ignore those dynamically generated
files. This practice was carried forward when the project switched to
the Meson build tool out of expediency. This change takes one step
towards fixing that undesirable practice by moving the generated conftab
source files into the build tree.

This change slightly simplifies the conf.sh script. It also introduces a
new helper script to convert a template file into a target file using
Python. Since we already depend on Python indirectly via our dependency
on Meson this isn't introducing a new dependency. Using it to translate
a template into a concrete file is preferable to the alternatives such
as using shell constructs like

    eval `cat some_file`

I don't think I need to explain why `eval` in that context is dangerous
and fragile. Similarly using `sed` is fragile because substitutions that
have embedded newlines require escaping the newline and potentially
other characters.

It also introduces a helper script to sort the unknown C preprocessor
symbols to deal with a bug in the macOS `grep` command.