Don't declare strlcat() unconditionally
Recent changes cause the unconditional declarations of `sprintf()`,
`strlcat()` and `strlcpy()` to be used (in a few files) after the relevant
system header that declares them has been imported. On macOS this causes
compilation errors due to differing declarations. So check if the system
provides those functions and only declare them if it doesn't.