Fix a use after free issue caused by _nv_unset function
_nv_unset function runs stkclose() on a stack in a while loop until
it's reference count is less than one and it can free the stack. But
the same stack is getting free'd by sh_freeup() function later. This
may cause a crash due to memory corruption. In ksh93u+ release,
stakdelete() macro was used to close the stack. This macro calls
stkclose() function and it decreases reference count for the stack by
only one. It should be later free'd by sh_freeup() function.

Related: #398