Revert `sh_open()` implementation to ksh93u
This is the first step in removing `pathdev()` and `pathopen()`. Those
functions were added to libast after the ksh93u release in order to provide
the `sh_open()` semantics to users of libast. That includes mimicing 3DFS
behavior on systems without that file system type. The new code has bugs
and is incomprehensible. Rather than trying to fix it revert to the ksh93u
implementation of `sh_open()` to not use those functions.

Partial fix for #510