Remove `pathopen()`
If the magic "PATH_DEV" support that simulates features of 3DFS (3D file
system) is disabled there isn't any need for `pathopen()` so replace it
with `openat()`.

This removes one unit test that depends on the simulated 3DFS behavior
and which does not pass using ksh93u.

Part of resolving #510