Fix svar_init()
This fixes the initialization of `struct Svars` by making the amount of
space allocated match the recorded `dsize`. Whether or not this is correct
is unclear because I don't understand the inclusion of `sizeof(int)`
in the calculation of `dsize`. But it does keep the `strncpy()` in
`sh_setsiginfo()` from modifying data past the end of the allocated buffer.

Fixes #556