Fix bug in sigalrm() setting shp->siginfo[SIGALRM]
The `sigalrm()` function has two bugs. First, it's copying the `ucontext_t`
parameter when it should be copying the `siginfo_t` parameter. Second,
it's incorrect to assign a `siginfo_t*` to `shp->siginfo[SIGALRM]` since
that is supposed to be a `struct Siginfo*`.

I found this while looking at a core dump on OpenBSD that had sigalrm()
on the stack. I don't think this bug is the cause of the core dump but
it's wrong and could cause a fault. The crux of the problem is that
shp->siginfo[SIGALRM] is a linked list of `struct siginfo_ll` but
`sigalrm()` was treating it like single `siginfo_t` element.