Disable one `pty` test on OpenBSD
There is one `pty` test that fails consistently on OpenBSD. I have zero
interest in figuring out why given that the pty command is obscure and thus
little used. This is the final test that fails consistently on OpenBSD. At
this juncture having `meson test` pass with zero errors (other than flakey
tests) on OpenBSD is more important than fixing this issue. This makes
it much easier to focus on problems involving the core ksh functionality.