Fix definition of SF_SYNC
This fixes the API/texcept.c API/tsync.c unit tests on FreeBSD.

On FreeBSD the sys/socket.h header contains `#define SF_SYNC 4`. Because
that header was included after the AST header that defines that symbol
we end up with the wrong value. Worse, because this is sensitive to the
headers have been included before sfhdr.h one module may end up with
the correct definition while a different module gets the wrong one (that
was the source of this bug). Moving all the system headers to the top of
sfhdr.h provides the needed consistency.

Partial fix for #577

P.S., This sort of header reordering needs to be done systematically
across the entire AST project.