Eliminate O_INTERCEPT
Now that we no longer have to deal with nested pathdev() calls we don't
need the O_INTERCEPT flag. Remove it.

I noticed while doing this another invalid constant definition of a symbol
that blindly assumes a particular bit mask is valid. Define O_SEARCH in
terms of O_PATH if the system provides it.