The generated lcgen.c file needs config_ast.h
I recently tried building on Cygwin and it failed to compile due to the
redefinition of `malloc()` and friends. This is due to how those
functions are implemented in Cygwin via functions in the system headers.
So ensure we include those system headers before we do things like
`#define malloc ast_malloc`.