Fix another pathdev() removal botch
The earlier changes that removed pathdev() introduced an invalid
transformation to openat(). But because the transformation was incorrect
it inadvertently produced the correct result. Looking more closely at
the code the pathdev(), and thus the openat() which replaced it, isn't
actually needed given that we're no longer supporting the 3DFS
emulation. So remove the invalid openat().