Enable unit tests on macOS
I eliminated all unit test failures on macOS some weeks ago. So it is
time to enable them in the Travis macOS environment.

Don't hadcode the value of CORE_COUNT.  Instead query the system and
default to one core if we can't figure out the actual number of CPU cores.

Part of resolving issue #515