Remove some lint introduced by header cleanup
The cleanup of the `#include` statements has introduced some lint in the
univlib.h header. I have no idea why that header is manipulating symbols
like `symlink` in that manner. If and when someone tries to build ksh on
a system that supports multiple universes (i.e., OS personalities) and
the build fails we can revisit this header.