Remove SHOPT_REGRESS
We haven't built with SHOPT_REGRESS defined true since the project was open
sourced. Looking at what this does it is not obvious the value justifies
the complexity. Quite the contrary. Given the availability of virtualized
OS environments it would be better to do the equivalent in a VM using
commands like `sudo` to actually create situations where euid != ruid.

Resolves #614