Lint cleanups in jobs.c
Oclint was pointing out that on macOS WIFCONTINUED() is a constant.
There is absolutely no reason for defining a private implementation in
that fashion as the macOS provided version works just fine. Too, it
should be available on every platform we support. Too, if this was
really needed it would also be needed for WIFSTOPPED.