Remove `_lib_sigvec`
Remove the `_lib_sigvec` feature symbol as `sigvec()` is an ancient,
obsolete, function that has been replaced by `sigaction()`.

Simplify the signal.c module.

Most importantly, both libast and ksh do `#define signal` to their own
private implementations that are not compatible. There are also some uses
of the libast signal manipulation functions in the ksh code that should
be using ksh provided equivalents. So modify the headers to ensure that
only one of the headers is included. Before this some ksh modules included
both fault.h and sig.h which meant the ksh module was using the libast
implementation when it intended to use the ksh implementation.

Resolves #630