Remove tests of SFIO emulated stdio functions
We dropped the SFIO emulation of the stdio API with commit 61c05a9. That
change overlooked a few unit tests that verified the correctness of that
emulation. Which hasn't been a problem until we enabled those tests and
tried to run them on OpenBSD which doesn't implement some things. For
example, `%S` as an alias for `%ls` in the swscanf() family of functions.
So drop those unit tests since all they're now doing is verifying the
stdio implementation provided by the system which isn't our responsibility.

Fixes #641