Remove SHOPT_ENV and the code predicated on it
Note that this also removes code predicated on `#ifdef _ENV_H` and similar
constructs because those depend on `SHOPT_ENV`. That code is broken and not
enabled in either the ksh93u+ release or any subsequent version. It appears
to be either an experiment or a work in progress that was abandoned.

Resolves #504