Simplify code and address Coverity Scan warnings
Coverity Scan was warning about `sh_open()` potentially accessing bytes
past the end of a pathname. Many, possibly all, of those warnings were
false positives. Nonetheless, the way the code is structured it could
theoretically access a byte past the end of the pathname string. Perhaps
more importantly the code is hard to read. It's not at all obvious it's
checking for "/dev/fd/" for example.